Build node mapping per thread
The single-threaded component of the huge node import is to build the
neoId -> graphId mapping, which runs in sequence over the loaded nodes
and puts them into a sparse array.

This changes, so that the sparse array is written to by the
importing/scanning threads and we can copy to the final data
page-by-page. In order to write into the Sparse array from multiple
threads, the page array is changed to an atomic array and
we lock when we have to create a new page.