Rely on static procedure list file
Rather than using a special compiler hook-in library to process
`@Procedure` annotations when compiling `:proc`, we now rely on a
checked-in static list.

It remains an open problem how exactly to make documentation updates
more dynamic to code changes, but we think using techniques that are
repeatable on the JAR is more sound than having to recompile the
sources.