Allow for deeper namespaces when delegation from algo.memrec
Before this commit, memrec would search for the pattern `algo.${name}.memrec`,
which would not match an `algo.graph.load.memrec` or other future algos that
have a deeper namespace. With this change, the user can provide `graph.load`
as algo name and we will find this one now.

Also, the available procedures in case of an error are now sorted by name.