Increase accuracy for community percentiles
We were already seeing wrong percentile counts for p100 on the scale of 10k nodes.
The Hdr Histogram is a probabilistic data structure and does not guarantee 100% accuracy.
The changed argument von 2 to 5 is the number of significant value digits, and if the recorded
range exceeds this number, previously recorded values will be merged into the same bucket
and we will see a loss in accuracy. The now used value of 5 is the highest that is allowed.

Co-authored-by: Martin Junghanns <martin.junghanns@neo4j.com>