Refactor procedure table and fix broken references
The procedure reference table was hard to read.
We simplify it to only contain the procedure name
and a short description.
The name is a link to the page that explains the
syntax and the parameters that the procedure accepts.
We introduce this for the product procedures, not yet
for the labs/experimental ones.

Co-Authored-By: Mats Rydberg <mats@neotechnology.com>