Initialize (reverse-)mappings in UF with number of communities
This is fine for the temporary forward mapping that we use in UF#init.
The forward mapping contains at most number_of_communities entries.
However, the reverse mapping grows by the number of new communities.
We don't know them upfront, it could happen that we need to increase
the mapping size which is bad ...

Co-authored-by: Max Kießling <max.kiessling@neotechnology.com>