Improve assertion for parallelStream test
We tested for stealCount == 0 but we cannot guarantee (on CI) that the
common pool is not being used by parallel threads (e.g. via parallel
test execution). We now check that the thread, that executes the lambda
on the stream has been spawned by the FJ pool and not the common pool.

Co-authored-by: Paul Horn <phorn@avantgarde-labs.de>