Move shared algo code to algo-common module
In order to split labs/product algo code, we need to handle a) actual
shared dependencies that don't belong in core, b) some not so nice code
entanglement. Both of these are solved by a new algo-common module, on
which the labs and algo modules depend.

The algo-common module shows us clearly some of the unhelpful coupling,
for example almost all of the pagerank code ends up in algo-common
because it cannot be cleanly split between product and labs.
Disentangling this is future work.

Co-Authored-By: Martin Junghanns <martin.junghanns@neotechnology.com>