Remove `nodeProperty` and `nodeWeight` from graph.load procedure
Both parameters were not documented and their usage was the
following: When set, the `nodeProperty` has been mapped to the
`seed` property key and `nodeWeight` has been mapped to the `weight`
property key (both defined by Label Propagation).  To make use of
those properties when running an algorithm, the user would need to
be aware of that internal mapping which is undocumented.

I believe we can safely remove those parameter and refer to the documented `nodeProperties` and `relationshipProperties.`