Iterate from node filter when computing
Adds a wrapper class to iterate over the set bits of a BitSet.
The wrapper class can yield a primitive stream of longs
based on the bits set in the NodeSet.
The iteration can start from the first set bit, or from an offset.
In NeighborhoodSimilarity, we use this to drive the outer (no offset)
and the inner (with offset) iteration to get node pairs for which to
compute similarity.

Also change the logging to not use an AtomicLong.