Compute in both directions when parallel topk
For single-threaded with topk we can compute in one direction and then
append the reverse of that result to the stream. With parallel streams,
this interferes with the partitioning and multiple threads will write to
the same topk priority queue, which is not safe. Instead we compute the
result in both directions. This doubles the number of computations, but
keeps the threads from interfering with each other.

Co-Authored-By: Martin Junghanns <martin.junghanns@neotechnology.com>