Remove unused centrality algorithm
BetweennessCentralitySuccessorBrandes was not exposed in a procedure,
and was annotated as unstable and deprecated because of deadlocks.
Removes that algorithm and the tests for it.

On the upside, we now don't have many identically named (and mostly
identical) tests for betweenness centrality.

On the downside, it turns out most of the testing was on the now removed
algorithm. The remaining algorithms are largely untested.