Use the actual average degree to define partition sizes in PageRank
PageRank was doing a very naïve estimation of using the fixed value 8 ([arguably a random number](https://xkcd.com/221/)) as the average degree to determine the approximate sizes of the partitions.
For the tests, that were assuming they would run in parallel, this "estimation" was too high, and every node got put into the same partition and no test actually ran in parallel.
When this was written, we didn't have the relationship count on the graph, but we do now.
We're now calculating the actual average degree and use that one, much win.

Co-Authored-By: Mats Rydberg <mats@neotechnology.com>
Signed-off-by: Paul Horn <paul.horn@neotechnology.com>