Move CypherMapWrapper to annotations
It doesn't feel right to live there, but we need it in a module that can
be completely build before the proc-processor module is being build,
which needs to be done before the core module can build.

Alternatively, we can introduce another module just for the wrapper that
lives between annotations and proc-processor.