Fix bug in NodeSimlarity graph building
The node count used to compute page size for building the result graph
was based on the number of nodes compared in the algorithm. It needs to
be based on the number of nodes in the base/input graph, or the page
size may be too small and index out of bounds will happen, either when
constructing or traversing the result graph.

Co-Authored-By: Jacob Sznajdman <breakanalysis@gmail.com>