Wrap calls to Neo4j's MapUtil
Add a "compatibility class" extending `org.neo4j.helpers.collection.MapUtil`.
We use Neo4j's MapUtil in many places, but the class lives in different
packages in Neo4j 3.5 and 4.0, which means it adds to the code gap that
we need to bridge to maintain compatibility with both Neo4j versions.
A no-op extension allows us to keep all the calling code the same, and
only update the compatibility class.