Align validation of invalid graph names
All catalog operations will now validate the graph name in the same way.
The one exception is `graph.list()` which allows the parameter to be
`null`, in which case it will list all graphs (no-arg assumed).

The tests all source invalid graph names from the same method.

Also added a result consumption to the assertError to handle lazy eval.