Add topK value to similarity proc tests
ANN expects a default topK of 3, and this was clearly the case in the
labs similarity algos. When setting a default value for the ported alpha
algos, however, several ported tests started failing. We did not remove
explicit topK config from the tests when we ported them, which suggests
that the topK default was somehow not applied when the tests ran. On the
other hand, ANN clearly expects the similarity algos to default to a
topK value of 3, so what gives?

We update the tests and defer for potential productization to set the
score right.