Use global aggregation for property mappings in CypherConfig builder
There were several places we could have made that change.
For example, we also compute the deduplication strategies
in the Huge and Cypher factory and could easily set the
potentially given global default if a single mapping has
a DEFAULT strategy. However, since we want to get rid of
this global default I opted for keeping those changes out
of the factory but made sure that whereever we set the
global strategy, the property mappings are getting updated
accordingly.