Separate thread pool config from algorithm config
- sets default concurrency to 4 for algorithms
- renames constants in Pools to be about thread pool size

Previously, the thread pool had the same size limitations as the
default number of cores to use. We now want to separate these in order
to have a static default of concurrency 4 for algorithms, while making
the thread pool as large as the system allows, in Enterprise deployment.
The thread pool is still limited in a Community deployment.

Co-authored-by: Paul Horn <paul.horn@neotechnology.com>