Fix invalid footnote references
An asciidoc cell in a table is basically it's own document as far as
Asciidoctor is concerned. Declaring a named footnote in one asciidoc
cell and referencing it in another therefore doesn't work. Luckily,
formatting these cells as asciidoc is pointless, so here we just stop
doing that. Easy peasy.