Disable same-thread-execution if task size is 1
The previous behaviour might lead to dead locks if within a
single procedure call, the transaction is used by the proc
thread and a newly spawned thread. The issue came up when
trying to write both, node properties and relationships in
the same procedure which caused a dead lock since relationships
were written in a separate thread which tried to acquire a
RW lock on a node which was altered by the proc thread and
not yet released.

Co-authored-by: Paul Horn <paul.horn@neotechnology.com>