Allow for properties to be missing
If we have a property mapping that resolves to a missing property
(i.e. propertyKeyId == -1), we now still "load" that property.
We won't hit the property store, but we allocate a property adjacency list
and fill it with the default values.
As per property validation, this is only allowed for COUNT aggregation with the '*' property.

Co-Authored-By: Veselin Nikolov <veselin.nikolov@neotechnology.com>