Subgraph doesn't reverse neighborhood functions
The subgraph builder reversed the neighborhood function list.
That means that the contract for the subgraph builder was, that caller
provides a list of neighborhood functions sorted in the reverse order to
what it really needs. That is not a great contract. A better contract
is that caller provides the list sorted in the correct order.

As a consequence, subgraph factory methods can be called with immutable
lists. This is nice in tests.