Replace use of FileUtils
We used FileUtils from both Neo4j and Apache Commons IO.
The former is not stable across versions, the latter (the version we
used) was File oriented, but Java and the rest of the world has moved on
to using nio Paths. Upgrading to commons-io 2.7 we can use PathUtil, and
move off of Neo4j's FileUtils altogether.

Co-Authored-By: Paul Horn <paul.horn@neotechnology.com>