Make access to label BitSet during import thread-safe
The code path is used by native and cypher loading as well as the GraphBuilder utility. The BitSet are created in a thread-safe manner, but the update of the BitSet itself is not thread-safe. This can only be triggered when importing in parallel, with labels, i.e. not for Cypher, but for native and GraphBuilder.