Compute nodes for summary graph sequentially
Parallel summary graph creation assigns new node ids in an arbitrary order. K1 computes its coloring based on that arbitrary order. The final result is not wrong (nodes end up in the same communities). However, the performance gain of parallelizing Louvain summary graph node creation is to low to justify adapting the test framework to be flexible with proc and doc results.

Co-authored-by: Max Kießling <max.kiessling@neotechnology.com>