Work around task order conflict in DocsArtifacts build
Task ordering still matters for the docs builds in the composite
artifacts pipeline. This time the reason for the failure is that the
docbook, pdf and html tasks do custom build log magic--if anything at
all is printed to stderr they throw an exception. When these tasks run
at the same time as some Java task, and the Java task prints a compiler
warning, the docs task fails.

By enforcing that these tasks run last, also within the docs-only
artifacts build, we can avoid this.

Also: we weren't actually running the tests in the :doc module. So let's
start doing that.