Inject AllocationTracker in BaseProc
We should store the AllocationTracker in a field, so that we don't risk
creating new ones on each access.
For the Neo4j memory-tracker-backed AT it doesn't matter so much, since
the backing memory tracker will be the same.
For our own in-memory implementaiton we would get non-shared tracking.

We cant have fields in proc classes except once marked with
`@Context` and injected. So we register an extension that let's us
inject the AllocationTracker this way.