Let test impl decide which proc gets registered
In many test bases we register procs for all modes. At the same time we
force the concrete tests to implement `getProcedureClazz`. Instead of
registering procs for all modes we can use this method to determine what
to register.

This also makes sure we don't call the wrong procedure mode in the test,
because that other mode won't be there.