Fix flaky CypherFactoryTest
The type of a property in the in-memory graph depends on the type of the
first value for that property found in the database.
Therefore, depending on the order in which the nodes are loaded, either
an actual value, or the value given in the coalesce() espression will be
used. But that resulted in different typing of the in-memory graph, and
therefore in different canonicalizations, causing the test to
occassionally fail.