Add LogisticLoss variable
This is computes cross entropy loss on top of logistic regression and
includes a direct connection from weights to loss.
It propagates gradient directly to weights without first propagating to
"predictions".
However it takes predictions as a parent so that predictions dont have
to computed both for forward and backward pass; by declaring predictions
as a parent, the computation will be cached in the computation context.

Co-Authored-By: Veselin Nikolov <veselin.nikolov@neotechnology.com>
Co-Authored-By: Jonatan Jäderberg <jonatan.jaderberg@neotechnology.com>