Remove flawed optimization from IdMap.withFilteredLabels
The optimization returns the original id map, if all nodes in the id map have a requested label. The resulting NodeFilteredGraph does a nested id lookup, which - in case of the optimization - is the same id map and the inner lookup returns the original (neo) id. That id is used to lookup the outer original id, which leads to an AIOOB exception.

Co-authored-by: Sören Reichardt <soren.reichardt@neotechnology.com>