Add failing test on RelationshipsBuilder
Writing relationships on a filtered graph does not map the node space correctly.
It writes relationships based on the mapped *and* filtered ids.
After adding the relationships to the graph store, we silently assume that the written ids relate to the unfiltered graph.
The test adds a number of offsetting nodes before the filtered label to show two different ways in how that bug manifests.

Co-Authored-By: Martin Junghanns <martin.junghanns@neotechnology.com>