Use multi-graph approach with UnionGraph to power Link Prediction
Instead of resolving GraphStore -> 2x Graph at AlgoFactory time,
we repackage the config and use the standard GraphStore -> UnionGraph resolution.
By doing it this way we know the Graph is a UnionGraph going into the algorithm,
where we use the multi-graph API to filter on relationship types.

This is not hack-free, but it is arguably more elegant than the AlgorithmFactory approach.
Also, it contains the hackyness to LinkPredictionTrainConfig only.