Replace TransientAdjacencyOffsets implementation with HLA wrapper
The prior implementation of TAO used a dynamic page size which depended on the node count and the concurrency. That page size has been used to access the inner page and the index in that page.

With the latest change to directly writing offsets into a HLA, the page size is fixed to 2^14 and _not_ depending on concurrency or node count. However, when creating the offsets we still forwarded the page size to the TAO, which then used it when computing page and array index. This lead to an AIOOB on large graphs with high concurrencies (basically any scenario that changed the default page size to somethin which is not 2^14).

This PR aligns the offset factory API with the degree factory API by directly consuming a HLA instead of long[][] which in turn simplifies the TAO implementation to be a simple wrapper around the HLA.

Co-authored-by: Max Kießling <max.kiessling@neotechnology.com>