Implement ArrayScaler to optimise memory access of array properties
Previously, we were accessing only one index of an array at a time during scaling computation.
Now, we will access all indices of the same array property in sequence before moving on to the next node.
This may be beneficial for memory access patterns.

Additionally, this commit removes the Supplier delegation used in ScaleProperties, which means we materialise each scaler immediately.

Co-authored-by: Florentin Dörre <florentin.dorre@neotechnology.com>