Add option to provide a local protoc installation
The protobuf team builds some protoc binaries and offers them for download via maven coordinates.
However, they only build x86 architecture binaries.
If one were to use, e.g., an M1 Macbook, which is an aarch64 architecture, the protobuf plugin fails to initialize.
Protobuf/protoc actually builds and works on aarch64, it's just not precompiled and provided.
Such unlucky folks on those machines could sift through the depths of the interwebs for instruction on how to build their own protoc, change the plugin definition to use the local installation and stash and unstash that change for months…
Or we could merge this PR and make some lifes easier 🙃