Replace transaction internal types with the transaction
Some types on the compat API surface are types that are per transaction, such as PageCacheCursorTracer.
Since those are likely subject to change (in fact, across 4.0-4.3-drop04 we have at least three cycles of refactoring around those types), we change the compat API to accept the KernelTransaction.
This allows each compat layer to only pull the data that they need.
The downside is that we have to at least provide the KernelTransaction everywhere.
In most places, we already have it available and the usages don't make sense without a transaction.
The only place where this was not the case is in Geri, where we use the PageCache, but do not have a transaction.
All usages before would pass a NULL PageCacheCursorTracer, so remove the transaction from that compat API completely and let the compat layer decide which particular NULL object is correct.
Though not strictly required, the MemoryTracker is also removed from the compat API surface.
We could remove some methods from the compat API after the change.