Allow creating only transaction-fed SecureTransactions
Prior to this commit, it was possible to create a SecureTransaction by passing a `null` Transaction which would elevate the security context to full access.
Many procedures constructed their exporters (classes that write to the db) this way.
To prevent this mistake, SecureTransaction can now only be created using a real Transaction.
This fixes several bugs where procedures were allowed to write to the database even when the RBAC rules would disallow it.

SecureTransaction currently wraps a Transaction and a GraphDatabaseAPI, but it needs only the API for its transaction handling.
It uses the Transaction only to copy its SecurityContext which is used for any additionally created transactions.

Co-authored-by: Lasse Westh-Nielsen <lassewesth@gmail.com>