Avoid capturing `this` in UnionNodeProperties
JOL fails on UnionNodeProperties in combination with JDK 15+.
UnionNodeProperties#valueProducer stores a lambda which captures the UnionNodeProperties instance.
This leads to the situation that JOL by default cannot compute the size of that "hidden class".
JOL has a system property, called "magicFieldOffset" which is "false" by default.
Enabling the property allows JOL to use some magic to resolve the size, but its usage is not recommended according to [JOL code docs](https://github.com/openjdk/jol/blob/master/jol-core/src/main/java/org/openjdk/jol/vm/HotspotUnsafe.java#L43-L55).
We therefore adapted to code to not capture `this`, but instead providing this on the call site to the lambda.

Co-authored-by: Adam Schill Collberg <adam.schill.collberg@protonmail.com>