Remove access to NeoStores from Cursor scanners
This allows us to remove the dependency on the GraphDatabaseAPI for the construction of the scanners.
It was used to get the size of the underlying store, which we now calculate from the GraphDimensions.
Since the size was only used for logging, it's not critical if the calculations are not 100% exact.

Co-Authored-By: Mats Rydberg <mats@neotechnology.com>