Use original node mapping to write similarity result
The source and target identifiers used to create similarity pairs refer to the node id space of the root id map (see RelationshipsBuilder#fromInternal). This leads to issues when writing the resulting similarity graph, since it's (potentially filtered) node id map has been used for original id lookup. The given source / target can exceed the internal id space of the filtered node mapping and lead to AIOOB (in the best case). In the worst case, the id exists, but maps to a different original id, therefore writing to the wrong nodes.