Add the ability to generate GraphStoreValidation
This commits makes the configuration processor able to generate a body
for a method annotated with `@Configuration.GraphStoreValidation`.
The body will consist of calls to all available methods that are
annotated with `@Configuraion.GraphStoreValidationCheck`.

A few checks are in place to mandate the shape of these methods, but no
exact validations can be made as GraphStore is not a class available to
the configuration processor (nor should it be). An advantage of this is
that this way of generating bodies for specific check-methods can be
abstracted to support also other types of checks that require an
injection.

Unfortunately, I've been unable to locate a way for Immutables to do
something similar to this, meaning these classes will not get the same
method bodies and thus will not carry out the graph store validation
checks. However, as our configuration processor is increasing in
sophistication perhaps we can stop using Immutables for the config
classes in the future.

Another shortcoming of that fact is that the method may not be abstract,
but must be declared with `default`. This is unfortunate, but not a big
problem.