Do progress tracking in Pregel algos
including PageRank, but PR is a special case. We can declare the
progress task tree in a way that fits the general Pregel task
begin/finish pattern--but only if we skip the initial degree computing
step. This step is problematic anyway, because the computation actually
happens in the algorithm factory, and not in `PR#compute`. This means
that the outmost task needs to be begun there, in the factory.
That's doable, if a bit weird--but there is no corresponding place to
finish the outmost task.
This is a clean place to commit, but we will have to do more work if we
want the degree computation step reflected in PR progress
tracking/logging.