Move identical configuration option to base interface
This didn't work before because the type of PropertyMappings is a generated type and it was not available at this point in time. Apparently, the graph create configs were compiled _after_ the PropertyMappings, so it happened to work there incidentally. I have not verified the compilation order though, this is just an educated guess.

For this two work now, we need two things:

1. An update to auto-common. The version we were using before had a bug where it crashed on the named-but-not-yet existing type, that bug is fixed in 1.x. It also looks like the update version has a bugfix for the situation when something fails to generate and then nothing runs anymore. Getting flooded with hundreds of unrelated errors, no more (possibly)!
2. We need to name the generated types in their fully qualified version so that javapoet is able to generate an import for this. The code generation itself is fine, javapoet does the right thing but with the type not even existing, it can't figure out that it has to import the type from somewhere.