Use progress tracker as much as possible
To use it fully we would need to define a task tree for loading upfront,
and that will require some additional work--to figure out the task
volumes for the various tasks, as well as which tasks will be performed
in a particular load (are we loading node properties this time? rel
properties?) So this is not a full port of loading to the progress
tracker infrastructure. But it is taking us a bit closer, which is nice.