Correctly count all ids in bit id map
This fixes a bug where nodes on the final page were not counted.
The bug triggers only when the internal size of the bit id map array is
a multiple of 64 (has all zeros in 6 lower-order bits) AND there are
bits set in (at least one of) the last 64^2 positions.

Co-authored-by: Sören Reichardt <soer3nreichardt@googlemail.com>