Really fix the bug
There are three cases, depending on the size of the bit id array:

1. the size is <= 64
==> no whole blocks; there is only a tail
2. the size is an integer multiple of 64
==> only whole blocks; there is no tail
3. otherwise
==> X whole blocks; a tail of size <= 63

The bug was originally that we did not respect case 2., but was mistaken
for case 3.
For case 2. we must make sure to count also the last page in the first
block-loop, as we will not count it in the loop that deals with case 3.