TaskProgressTracker doesn't require ProgressLogger
Add progress tracker constructor that also takes a log and concurrency,
and that creates the BatchingProgressLogger itself--sparing the caller
from knowing about it.

Port most usages where a BPL was created just in order to create the
tracker, to use the new constructor.

Remains to touch the usage in AlgorithmFactory--and
`TestProgressLogger`s--if we want to remove the old constructor
altogether.