Guard against concurrent updates
Since we read the graph size in one transaction and load in multiple
different transactions, any new data that is being added during loading
will show up while scanning, but would not ne accounted for when
sizing the data structures used for loading.

The node loading part now only accepts nodes that are within the
calculated capacity that we have available.