Defer allocation of nodeCount long collection during relationship loading
We buffer relationships during loading and only write the adjacency list when we run the flushTasks.
We only need the offsets/degrees during this flush, but we allocate them before we import any relationships.
This commit defers the allocation of those types so that we can also defer the point in time when we request the number of nodes.
This allows us to have a node mapping that will be gradually built while we import/buffer relationships.