Include :sysinfo already in :proc
This means it's not a leaf project after all.
To make it a leaf project we would need to move `ListProcTest`, or else
have a secret convention that we have these additional funcs/procs that
aren't covered by it.
This way we should hopefully get some improved build times for the
various proc projects.
But fully pushing out the build properties bits requires a bit more
work.