Fix an issue in the SimilarityGraphBuilder
For KNN the SimilarityGraphBuilder was called with the NodeMapping
of a potentially filtered Graph. This mapping would then be used to
create the new relationships as well as the SimilarityGraph.
This leads to a problem, because RelationshipsBuilder.addFromInteral
calls `toRootNodeId` on the source and target nodes. This means that
the actual relationship ids are not in the same id space as the
NodeMapping used to create it.
To fix this, the SimilarityGraphBuilder now receives two node mappings,
the one from the filtered graph as well as the root mapping, which is
used as the id map for the SimilarityGraph.

Co-Authored-By: Ioannis Panagiotas <ioannis.panagiotas@neotechnology.com>