Use DefaultFileSystemAbstraction in GraphStoreExportProcTest
The ephemeral file system changed the semantics of deleting files to throw if the specified file does not exist. The PBM however does not perform this check and will throw an error. To circumvent this we have to use a real file system with a temp directory (which was created before anyways).

Co-authored-by: Adam Schill Collberg <adam.schill.collberg@protonmail.com>