Forbid using Immutable configs in main scope
Config classes are generated using our code-generating annotation processor.
The implementation classes typically follow a `*Impl` naming pattern.
For convenience, we almost always double-up using also a `@ValueClass`
annotation to get an `immutables`-generated class as well.
The latter one comes with some nice builders.

However, this creates a situation where we are using different config
implementation classes in various parts of the system. This is a source
of bugs that can be very subtle, as the differences between the
generated classes are not entirely clear.

This rule attempts to stop us from using the `immutables`-generated
configs in `main` scope; using them in tests is still OK.