Depend on packaging docs sources earlier
IntelliJ doesn't run tests via the test task. It therefore wouldn't pick
up the dependency on packaging the docs sources for doc tests. By
pulling the dependency earlier, to processTestResources which is used
when IntelliJ runs tests, we get the docs sources provided whether we
run doc tests with Gradle or IJ

Co-Authored-By: Florentin Dörre <florentin.dorre@neotechnology.com>