Use loopback address instead of localhost
For some reason, java resolved localhost to a wrong IP.
I suspect it is the cached IP for the WiFi interface but
I am connected via ethernet and have WiFi disabled.

Using loopback seems to work fine and does not require an
actual DNS lookup to figure out where localhost actually is.

Co-Authored-By: Sören Reichardt <soren.reichardt@neotechnology.com>