Compute sign inside SignedProbabilities and replace 0 probabilities
Since random forest may produce a probability of 0, we replace such
values with 1e-100 to retain the sign (positivity vs negativity of LP
eddges).
Adding the sign as an argument to SignedProbabilities::add makes it
impossible to add a 0 value with ambiguos sign.
If the logic was at the call-site, then future call-sites would be
unprotected from such faults and would have to know about the 1e-100
constant that replaces 0.0.

Co-Authored-By: Mats Rydberg <mats.rydberg@neotechnology.com>