adding seeding to target node filtering
so that we can almost always ensure presenting users with k results. of
course, data might be poor and give poor results - you would use this
functionality as a last resort. also, there is no guarantee that there
even are k target nodes, so that is a theoretical hole.

but imagined use is, you seed with arbitrary target nodes and in the
course of KNN running, you find k better results anyway.