Delay logging of Sysout during proxy initialization
Neo4j docker – for reasons relevant to them - runs `neo4j console --dry-run`
at some point which is supposed to print (to stdout) a list of JVM flags
or a command line to execute.

This dry run also loads and verifies settings, but not kernel extensions.
Since our SettingsDeclarations involve a proxy layer to define settings,
we produce log output that we print to stdout, which confuses the later
interpretation of the output of `neo4j console --dry-run`.

With this PR, we buffer log messages internally and dump them to an actual
Log once we have one available.

Co-Authored-By: Jonatan Jäderberg <jonatan.jaderberg@neotechnology.com>
Co-Authored-By: Anton Klarén <anton.klaren@neotechnology.com>