Use single-threaded Map for data and gradient
The ComputationContext is instantiated per batch and thus only accessed by a single thread.
Therefore we dont need to have concurrent datastructures.

Maybe in the past we thought of executing multiple independent variables in a computation graph in parallel. However, until we dont do this, lets avoid the significant overhead for smaller computation graphs.