Downcast to `NodeEntity` to get stable signature
Neo4j 4 and 5 return a different type for the method

org.neo4j.graphdb.Node#getRelationships()

This causes a bytecode error, because the compiler records the exact
type of the method known at compile time.
There is no such one method that is the same across the Neo4j versions.

Co-authored-by: Adam Schill Collberg <adam.schill.collberg@protonmail.com>