Move Neo4jLayout.of behind compat layer
This is one of those sneaky compat issues.
In 4.x the signature is `Neo4jLayout.of(org.neo4j.configuration.Config)`
In 5.x the signature is `Neo4jLayout#of(org.neo4j.graphdb.config.Configuration)`.
`Config` is a concrete class implementing the `Configuration` interface, so 5.x moved to a more abstract API.
Since we always pass a `Config`, this is a source-compatible change.
The Java Bytecode however will contain the concrete type in the signature, so it tries to link and call a method expecting a `Config` but now it only finds one accepting a `Configuration`.