Support reading node/rel/graph properties with long names
When the name is long, the CsvEncoder will quote the string.
The default limit is 24, which we are not configuring explicitly.
The quoting itself is not a problem if we use the same library to read
the header files as we use for writing them. This commit achieves that.

An alternative solution could be to avoid all quoting, but I fear the
consequences of that for weirdly named properties with potential control
characters, whitespace or other "unexpected" characters.