Improve cursor caching in CSRCompositeRelationshipIterator
In the previous code, we used `AdjacencyCursor.empty()` which will not get reused if used in a compressed / uncompressed list as it's not not a compressed / uncompressed cursor.
Instead a new cursor will be instantiated at the first time.
Also, we initialized property cursors, but did not use them for reuse, which is a performance hit.

We can not reuse the same adjacency / property cursor for forward and inverse topology, as the raw adjacency cursor wraps the underlying pages, which are different between forward and inverse.