Eliminate race when mapping intermediate ids
We have a TOCTOU situation in the LazyIdMapBuilder.
We check if a node is already mapped and if not, we add it to the
nodesBuilder. This also works as a deduplication mechanism.
When we add the same node from multiple threads, those two threads can
both observe the node to be missing, and both will add the node.
This will overwrite the intermediate mapping for that node for one
thread, but we will insert the intermediate node ids from both threads into
the final id map. The missing mapping leads to a sparse HLA for the
originalId mapping, which we never expect to be sparse. This leads to an
originalId of 0 for those lost nodes and this can result in all kinds of
errors, e.g. a NOT_FOUND later on.

Co-Authored-By: Martin Junghanns <martin.junghanns@neotechnology.com>