Using the graph store DB name when exporting with Arrow
This is a fix: instead of using the name of the DB that is providing the context, we use the name of the DB
in the graph store to perform write backs. The graph store saves the name of the DB from which the graph
was projected.

This commit fixes the node properties and relationship exporters