Use rootIdMap for id resolving in NodeProperty builder
We use the root id map to access the innermost id mapping layer.
For all id maps except HighLimit, this is the same instance.
For the HighLimit id map, the root id map is the mapping from
intermediate to mapped ids. Since we use the intermediate ids
to write properties, we need to use that mapping to correctly
resolve the values.

It is somewhat hacky, I admit ...