Remove capacity check in NodesBatchBuffer#add
The check had been previously introduced to avoid the situation
where nodes are added to the database while running a native
projection leading to a potential AIOOB on the pre-sized data
structures.
This check is still in the #offer() method which is the one that
is being called on from the node scanners. The #add() method is
called from the NodesBuilder (construction) code path where this
constraint is not necessary.
Removing this check allows us to work nicely with high limit id map
where the external highest id is not the one we use for sizing the
internal data structures.