Validate input constraints when HighLimit is requested
When HighLimit is used, we require the node count to properly size
the label bit sets. Alternatively, we could just let them grow which
is a bit more runtime overhead.

When HighLimit is used, we cannot use deduplication as this requires
an additional HABS which might fail on creation due to size limitations.