Replace nested lambdas with anonymous classes
GraphInfo depends on JOL to compute the memory usage of a graph.
It does so by walking the object graph and computing the size of
the individual nodes.
In these wrappers, we store the method handle to array::get as a
lambda field, which is not possible to walk/measure by JOL.

Co-Authored-By: Paul Horn <paul.horn@neotechnology.com>