Relaxing procedure name extraction in docs
The ProcedureNameExtractor has this regex to verify procedure names in documentation:
(\\sgds\\.)(\\w+\\.)+(\\w+)
However this doesn’t accept gds.listProgress (that was detiered from gds.beta.listProgress)

We already have other procedures like gds.list but it seems to me we don’t fail on these because we don’t test their doc pages.

Replacing the + on the middle word in the regex to * fixes the problem. So this works:
(\\sgds\\.)(\\w+\\.)*(\\w+)