Remove the need for a CompatAccessMode
By deleting a few unit tests, we no longer need the CompatAccessMode.
Those unit tests were originally testing the differences between a
licensed GDS and OpenGDS by ensuring that only the licensed GDS respects
any RBAC rules. Since we removed that distinction from GDS EE and also
respect RBAC in OpenGDS, the tests are now doing the same thing.

Furthermore, we were testing by using APIs on the transaction that we
are not using anywhere in the product with implementations of the
AccessMode that are never being used outside of this test, so those
implementation are only relevant for the test and don't give confidence
that GDS generally respects RBAC. We do have an integration that for
that.

Co-Authored-By: Jonatan Jäderberg <jonatan.jaderberg@gmail.com>